+++ lib/chefspec/server_methods.rb
@@ -10,16 +10,7 @@ module ChefSpec
     # @return [ChefZero::Server]
     #
     def server
-      @server ||= ChefZero::Server.new(
-        # Set the log level from RSpec, defaulting to warn
-        log_level:  RSpec.configuration.log_level || :warn,
-
-        # Set a random port so ChefSpec may be run in multiple contexts
-        port: port,
-
-        # Set the data store
-        data_store: data_store(RSpec.configuration.server_runner_data_store),
-      )
+      ChefSpec::ZeroServer.server
     end
 
     #
@@ -160,7 +151,7 @@ module ChefSpec
     #   to the server
     #
     def load_data(name, key, data = {})
-      @server.load_data({ key => { name => data } })
+      server.load_data({ key => { name => data } })
     end
 
     #
@@ -170,31 +161,10 @@ module ChefSpec
       args.unshift('organizations', 'chef')
 
       if args.size == 3
-        @server.data_store.list(args)
+        server.data_store.list(args)
       else
-        @server.data_store.get(args)
+        server.data_store.get(args)
       end
-    end
-
-    #
-    # Generate the DataStore object to be passed in to the ChefZero::Server object
-    #
-    def data_store(option)
-      require "chef_zero/data_store/default_facade"
-
-      store = case option
-              when :in_memory
-                require "chef_zero/data_store/memory_store_v2"
-                ChefZero::DataStore::MemoryStoreV2.new
-              when :on_disk
-                require "tmpdir"
-                require "chef_zero/data_store/raw_file_store"
-                ChefZero::DataStore::RawFileStore.new(Dir.mktmpdir)
-              else
-                raise ArgumentError, ":#{option} is not a valid server_runner_data_store option. Please use either :in_memory or :on_disk."
-              end
-
-      ChefZero::DataStore::DefaultFacade.new(store, "chef", true)
     end
   end
 end
